Moved all methods to modules (#460)
* Moved all methods to modules

* removed old mongoDB file
